[MRM-1398],[MRM-1401],[MRM-1403] submitted by Patti Arachchige Eshan Sudharaka o fixed repositories page o do not merge snapshot artifacts o fixed problem when merging artifacts with classifier
git-svn-id: https://svn.apache.org/repos/asf/archiva/branches/archiva-MRM-980@983343 13f79535-47bb-0310-9956-ffa450edef68

+++ archiva-modules/plugins/stage-repository-merge/src/main/java/org/apache/archiva/stagerepository/merge/Maven2RepositoryMerger.java
@@ -173,7 +173,7 @@ public class Maven2RepositoryMerger
         File targetPomFile = new File( targetRepoPath, artifactPath.substring( 0, lastIndex ) + "/" +
             artifactPath.substring( lastIndex + 1 ).substring( 0, last ) + ".pom" );
 
-        if ( !targetPomFile.exists() )
+        if ( !targetPomFile.exists() && sourcePomFile.exists() )
         {
             copyFile( sourcePomFile, targetPomFile );
         }
@@ -358,15 +358,12 @@ public class Maven2RepositoryMerger
 
         if ( ( sourceArtifact.getNamespace().equals( targetArtifact.getNamespace() ) ) &&
             ( sourceArtifact.getProject().equals( targetArtifact.getProject() ) ) &&
-            ( sourceArtifact.getVersion().equals( targetArtifact.getVersion() ) ) &&
             ( sourceArtifact.getId().equals( targetArtifact.getId() ) ) &&
             ( sourceArtifact.getProjectVersion().equals( targetArtifact.getProjectVersion() ) ) )
 
         {
-            if ( sourceArtifact.getId().equals( targetArtifact.getId() ) )
-            {
-                isSame = true;
-            }
+            isSame = true;
+
         }
 
         return isSame;
